Create ColorTrap.Java
This code was developed alone. This was a project done for an event driven programming class. The point of the game is to take trick the mind in assuming the color it reads is the answer, but the answer is actually the color of the word itself. The color choices given alternate location and color after every choice as well. There is a 15 second timer in place. Once the time is up, a total score is given. The user has the option to restart from the screen displaying the score.