Create Calculator.java
This was a group project done by myself and 4 others. I tried initializing the code I was primarily responsible for with the LV initials although my hands were all over the place in this. I would of personally done this with a stack but this the route the group chose to take. This was good learning experience in terms of working with other minds, styles of coding, and overall approaches to work.