AnimateGameOfLife.java
An individual project. This file is the file that gives animates and displays the pane and scene for the Conway's Game of Life. This file calls on the randomizeInitialGrid() method to make sure the game of life always takes on a new setting. The updateGrid() method returns an updated grid after the rules of life have been checked and changes have been made. This runs for 100 cycles.