GameOfLife.java
This file is a follow up on AnimateGameOfLife.java, GameOfLife.java is the logic behind it. In this file, we initialize the grid, keep track of the neighbors of the cells containing 'life'. The neighbor count is then used to update the grid after the rules of life have been applied.